b160

+++ advanced/phoneme_advanced-mr2-dev/cdc/src/share/foundation/classes/sun/security/util/DerOutputStream.java
@@ -1,7 +1,7 @@
 /*
  * @(#)DerOutputStream.java	1.46 06/10/10
  *
- * Copyright  1990-2006 Sun Microsystems, Inc. All Rights Reserved.  
+ * Copyright  1990-2008 Sun Microsystems, Inc. All Rights Reserved.  
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER  
  *   
  * This program is free software; you can redistribute it and/or  
@@ -198,12 +198,12 @@ extends ByteArrayOutputStream implements DerEncoder {
 	// Reduce them to the least number of bytes needed to
 	// represent this int
 	
-        if (bytes[0] == 0xff) {
+        if (bytes[0] == (byte)0xff) {
 	    
 	    // Eliminate redundant 0xff 
 
             for (int j = 0; j < 3; j++) {
-                if ((bytes[j] == 0xff) && 
+                if ((bytes[j] == (byte)0xff) && 
 		    ((bytes[j+1] & 0x80) == 0x80)) 
                     start++;
                 else
